chore: add mock API (#4)
* chore: Update npm dependency to latest stable version

This commit updates the `eslint` dependency in the `package.json` file to version `8.57.0`. The previous version was `^8.57.0`, which allowed for any compatible version within the `8.57.x` range. By specifying the exact version, we ensure consistency and avoid potential compatibility issues.

Note: This commit message is based on the provided code changes and recent repository commits.

* chore: Remove unused API routes

This commit removes the following unused API routes:
- `src/app/api/post/[username]/[slug]/route.ts`
- `src/app/api/post/[username]/drafts/[slug]/route.ts`
- `src/app/api/post/[username]/drafts/route.ts`
- `src/app/api/post/[username]/route.ts`
- `src/app/api/post/route.ts`
- `src/app/api/user/route.ts`

These routes are no longer needed and can be safely deleted.

Note: This commit message is based on the provided code changes and recent repository commits.

* chore: add mock api